some fixes for 64-bit and overflowing

+++ apps/output_cpp/gm_graph/inc/gm_graph_typedef.h
@@ -1,12 +1,12 @@
 #ifndef GM_GRAPH_TYPEDEF_H
 #define GM_GRAPH_TYPEDEF_H
 
-#ifdef GM_NODE64
+#ifdef   GM_NODE64
 typedef int64_t edge_t;
 typedef int64_t node_t;
-#elif  GM_EDGE64
-typedef int32_t edge_t;
-typedef int64_t node_t;
+#elifdef GM_EDGE64
+typedef int64_t edge_t;
+typedef int32_t node_t;
 #else
 typedef int32_t edge_t;
 typedef int32_t node_t;
